+++ lib/presentation/router/app_router.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:swickky/presentation/screens/home_screen.dart';
 
 import '../../core/exceptions/route_exception.dart';
 import '../screens/splash_screen.dart';
@@ -14,6 +15,10 @@ class AppRouter {
       case splash:
         return MaterialPageRoute(
           builder: (_) => const SplashScreen(),
+        );
+      case home:
+        return MaterialPageRoute(
+          builder: (_) => const HomeScreen(),
         );
       default:
         throw const RouteException('Route not found!');